removed commented code, reorganized files
