+++ src/components/Catalog/Catalog.tsx
@@ -7,7 +7,7 @@ import { useDataParams } from "../../store/DataParamsContext";
 import Header from "../Layout/Header";
 import Banner from "./Banner";
 import Variables from "./Variables";
-import { TabMenuLabels } from "../../constants/time-series";
+import { TabMenuLabels } from "../../constants/ui";
 
 const Catalog: React.FC = () => {
   const { setVariable } = useDataParams();
