+++ src/constants/time-series.ts
@@ -5,11 +5,3 @@ export const DefaultParams = {
   BEGIN_TIME: "2019-01-01T00:00:00",
   END_TIME: "2020-01-01T00:00:00"
 };
-
-
-export const TabMenuLabels = {
-  CATALOG: "Catalog",
-  LOCATION: "Location",
-  DATE: "Date",
-  VISUALS: "Visuals",
-};
