+++ src/constants/ui.ts
@@ -1,0 +1,6 @@
+export const TabMenuLabels = {
+  CATALOG: "Catalog",
+  LOCATION: "Location",
+  DATE: "Date",
+  VISUALS: "Visuals",
+};
